requirements.txt
# Requires Python 3.10+
Flask==3.0.3; python_version >= "3.10"
tensorflow>=2.16.0,<3; python_version >= "3.10"
keras>=3.0.0,<4; python_version >= "3.10"
numpy==1.26.4; python_version >= "3.10"
Pillow==10.4.0; python_version >= "3.10"
Werkzeug==3.0.3; python_version >= "3.10"
gunicorn==22.0.0; python_version >= "3.10"
